build system and third-party library
Okay, let's start at the start. We'll need the development kit and low-level libraries from devkitpro. In addition, I'll be using 0xtob's 'libntxm' for the music part. I'll pull it all as is, maybe we'll focus on that later.

Everything else here is just there to ensure builds happen properly, when invoking 'make' at the root of the repository.

`make` can automate some recipes to build stuff, like 'what to do when a you want to turn a .cpp source into a .o compiled file. because we're using cross-compiling, many of these rules were redefined in devkitarm (the compiler suit from devkitpro project). I imported them as //common.mk// to be able to tweak them.
(more in http://sylvainhb.blogspot.be/search/label/devkitpro)

Now, an NDS binary runs on a dual-processor system, but it is an asymetric system, with two different processor accessing two different programs: the arm9 and the arm7. The arm9
runs our main program, and the arm7 has access to sound chip and all those little things that read buttons, touchscreen coordinates and ... WiFi. To get our program access those features, we need to have a relay (slave) program running on the ARM7, receiving commands and launching the appropriate suppport code.
This bridging code is to be found in ppp7/ subdirectory.
(more in http://sylvainhb.blogspot.be/search/label/arm7)