a first demo application. Not much to see here, but it shows how to get data from the .nds file, and how we setup the error functions to catch exceptions.
The root of our application is the MetaWindow.
but the code doesn't run very well just like that. Seems like we need to import more

the demo embeds a music file (sndtrk.xm) that will be played when we'll press the A button. To do so, I'm using the EFS (embedded file system) library from NODA, integrated to the 'ppp9' support library. The NDS image needs then to be processed by a tool (efs.exe under MS-systems), that should be installed at $EFSTOOL makefile variable.