Base features of the GUI Engine
This is not really for games, mostly for editors, but some part are used to manage macro-states of the game anyway.

The roles of the Engine are:

- prepare the video hardware
  - there are many possible configurations of the video memory. We give them symbolic names here as an attempt to be more coder-friendly.
- register Windows, that are context for UI interactions and containers of Widgets
- dispatch keys/touchscreen events to the appropriate Widgets

just one thing: we need to make sure we'll invoke "sync()" on a regular basis. This will suspend execution until the hardware is ready to process one more frame of video (and this is critical for emulators to process our programs right

some incoming components (animation, interrupts and resource management) will need to have their runtime functions invoked correctly. The proper places are identified with HOOKs.