SpriteSet: your VRAM content manager A spriteset is a collection of graphic blocks that can be used to build background layers as well as moving elements. It can be used to fill the video memory with contents saved into a .spr file, as used by GEDS editors.
The internal details of the .spr file format is inspired by Amiga IFF files: a readable tag, a size, and then the content per se. There are different section e.g. for color maps (palettes), pixels values (tiles) and logical-block-to-physical-offsets (pages). This is all defined in SprFile.h

The SpriteSet is the control class for a .spr file loaded in memory. The SpriteRam class let you control where the graphic data will be loaded (sub-screen, main screen, background or sprites memory).

NOTE: libgeds split the functions of SpriteSet into multiple source files depending of whether they are useful for game engine or for editors. Only the read-only functions for game engine are imported here. See 7bbeb8322ea1 for the full set.