demo tutorial updated: loading background pixels
This demo will print the graphics contained in bg.spr on the top screen. To do that, we must load the data in the video memory of the chip in charge of the top screen. There are already some ASCII font in this memory and GuiEngine has reserved some areas to be the 'screen maps' for some layers. The sprite data will be loaded after all this, leaving room for 512 16-color characters before, as defined by `WIDGETS_CHARSET(512)`.

BG_PALETTE is another area of the video memory, where we program a set of 256 colors to be used by the background layers. the SpriteSet will take care of filling it with the data contained in bg.spr on `Load()` call.

Now, we still have two things to do: tell the video chip to show some layer, and decide what this layer should show. Every 8x8 square of graphics (dubbed a 'tile') we have loaded has been assigned a number (starting with tile number 256 when we use `BG_TILE_BASE(0)` and `BG_COLOR_256` because 256-color data takes twice the room of 16-color data).

GuiEngine.h provide some symbolic names for locations in VRAM where the maps are expected. This is the case with WIDGETS_BACKGROUND, which is a 32x32 array of 16-bit values where we can write tile numbers that we want to see appearing. To tell the chip we want this map showing on screen, all we still have to do is
- ensure the bit controlling one of the layers (here layer 2) is set in REG_DISPCNT, the master control register for top-screen display
- ensure the corresponding layer's control register programs the right VRAM slot (GEBGROUND)
- put some values between 256 and 256+last_tile in WIDGETS_BACKGROUND[row * 32 + col]

You'll notice the [o];[o+1];[o+32];[o+33] pattern in the offsets when filling the map. This is because a 16x16 block in the sprite editor has been assigned 4 consecutive tiles. If we just wrote increasing numbers in the WIDGETS_BACKGROUND[] array, we'd see the bottom half of the block on the right fo the top side, instead of being under it.