Resources and OAM management A page contains all the information required to program the video hardware : the shape, the sizes. We will need some easy way to manipulate the bitfields of the Object Attributes Memory -- the Nintendo name for sprite control blocks. This is `gSpriteEntry`, that comes from libnds.
There is a limited number of OAM entries (128 per screen), so it will be useful to allow our programs to decide who manages which entry. This (and similar features for other graphic resources) is delegated to the Resource object. Code that wants to allocate graphical resources should just derive from UsingResources and invoke `gResources->allocate()`.

We add functions to perform initializations (`setOAM()`) and updates (`changeOAM()`) of OAM entries straight through the SpritePage object.