Sprites / OAM runtime support
We usually don't modify OAM entries directly. In order to have flicker-less game,  we leave the real OAM entries untouched while the graphic chip of the NDS is updating the screen.

For this, we have UsingSprites::sprites[] array, that will serve as a 'shadow copy' of the real entries. We can update them anytime we like and Engine::sync() will take care of copying them when suitable (that is, during the pause -- known as vertical blanking -- between last line redraw of a frame and first line redraw of the next frame).