Tutorial #3: our first sprite
Sprites are likely the most important element of video games. The term stands for freely movable objects on screen, usually including player's avatar, opponents, but also some pickups and effects.

Sprites need graphical data, just like background layers, and these data are -- again -- made of tiles. We'll mostly use 16x16 blocks made of four tiles here, but NDS sprites can actually be from 8x8 to 64x64 (with some restriction). The task of updating hardware resources so that sprites show up on screen are delegated to SpritePage (that tells which graphics to use, what settings) and GuiEngine::sync() that will transparently update all the sprites coordinates at the best time given hardware constraints (i.e. avoid flickering).

The toy class 'Hero' shows all we need to have a freely moving object:
- remember current coordinates,
- have a SpritePage that knows how to update some OAM entry (the Object Attribute Memory, Nintendo's term for sprite control block)
- have allocated one of these entry and remember which one

By saying that this class is "UsingSprites", we gain access to the temporary location (`sprites[]`) that the engine will update during sync() calls, but also to the resource allocator `gResources' that can give us an OAM slot.

Loading the graphics data for the sprites behaves much like we did for background graphics in [[https://github.com/PypeBros/libgeds-tutorials/commit/67fdf2765a18f246ea74b822739017bab677afad|the previous tutorial]], except that this time, we must ensure that the SpriteSet and SpriteRam objects will live as long as the Hero that will need one of its pages. So we make them members of MetaWindow instead of temporary variables.

All we need now to give user the control over the sprite position is some test on the NDS inputs (the DPAD) and encode the corresponding move() function calls. Do not expect something smooth here: we are still in the 'editors user interface' and we receive an 'event' about the keys only when they have changed, and more specifically, only when some button get pressed. We shall fix that next time with the Animators system.