This is what it takes to have play() function called at desired period
Animator object really allows you to think about your game as lots of little objects living their life in a shared world. You don't need to worry what code to call when: the 'play()' functions will be invoked at the right time within the prepare-sync-update loop of the game engine.

The real thing is implemented into AnimationScheduler, that has multiple lists and take decisions or add/remove things as requested. From the external world, it all appears like being an Animator makes you a bit magic, but it is really because Animator::Animate() knows how to delegate to the real AnimationScheduler.

Some things in the rest of the code shouldn't happen while we are running the Animators code, such as we shouldn't change the current 'Window' (which might represent the world in which objects live). So Engine::setWindow() now has a mechanism to delegate the actual action past the Animate() step.