Tutorial #4: use Animator to move the Hero smoothly.
Explicitly calling move() over every objects will make the code of the game pretty complex as it evolves. How would you deal with bullets ? how would you deal with rooms that have different amount of objects ?

One step towards solving that is to have an 'animations engine' that can let you think of your code made of multiple, independent units -- the Animators -- that each take care of a little part of the game (one character, one animated block, one effect, etc.) extending Animator is the way to go.

Hero will thus be an Animator (of the OAM we move on screen). It should thus implement a `play()` method that will be called every time the Engine finds it right. We could tell how many frames to wait between two calls to play(), but we asked here (through constructor arguments to Animator(0)) to be called once per frame.

We no longer need a move() method, thus. Instead, the animator will check the DPAD status with libnds' `keysHeld()` (that were already pressed last time) and `keysDown()` (that were only pressed this frame) and define the new speeds locally.

returning `QUEUE` at the end of play() will tell the Engine the Hero survived the frame and should be called next frame too.

Finally, we need to tell the Engine when our Hero is ready to see its play() function called. We can do that here right after creating it, and that's the job of regAnim (register the animator).