the expression evaluation system
(see README.md for a description of the expression syntax and semantics)
expressions are parsed from ASCII to internal representation, and only the internal 'bytecode' is used during the game. A separate array is used to store immediate constants, so that we can stick to a simple one-byte-per instruction. An internal "use constant nr %%" command is written to link the opcode to the immediates array.

Expressions are normally stored in the 'state machines' tank, but we might need them freely-allocated later, for conditional statements or other evaluate-during-load instructions for the scripts. That's why they are templates, indicating how their allocation should work.