[tutorial] and at last, let's have our character move again. it is simply a matter of getting the compiled controllers linked against other components of our demo and indicating which controllers we use, and in what order.
At everyframe, 'xdad' will thus read the dpad information, store it in gobvar #2, then use the left/right direction to update the horizontal speed, and finally use the up/down direction information to update the vertical speed.

The 'to 512' part tells the limit at which we stop increasing the speed (whatever the direction). The unit is 1/256th of pixel per frame, so here the character can move up to 2pixels / frame along each axis.

see [[http://sylvainhb.blogspot.be/2018/04/to-ar-or-not-to-ar.html]] for more details about the need for `OFILES += $(libplugins_OFILES)`.