tutorial: the background now comes from a level-edited .map file
Map files are normally created with [[https://sylvainhb.blogspot.com/p/level-editor-for-ds.html|Level Editor for DS]]. They are an extension to .spr files that can store one or more "layers", that is, grids of tile selection.

libgeds/pixdox/mapedit-all.png shows you the different layers contained in demo.map: the foreground (that will hide sprites), the background (that will be hidden by sprites) and the 'meta' layer containing collision properties of every tile in the level.

The script use the name 'bg0' for the top-most layer, 'bg1' for the bg0-pinned background and has 'bg2' available for parallax scrolling. Once the .map file is loaded and the main layer is set up, a second command is required to prepare the pinned background in video memory. 'bg0.map:1' identify the "second layer" of the map loaded into bg0.

We need layers control registers to be tweaked so that the hardware is in the state that GameScript expects. You'll note that we use 2 'map index' per layer (that is, 2 x 32 x 32 tiles). That's because we want to be able to scroll and thus allocate two screens worth of tile map in video memory so that we can have some off-screen data to update while on-screen data is untouched.