world_coords_t vs. world_pixel_t for object coordinates
If we want to be able to have game entities move at any angle, we need them to be allowed to have sub-pixel speeds and to remember sub-pixel positions so that applying 1/4 pixel speed 4 times lead to one pixel move.

This is the job of world_coords_t while world_pixel_t express things coordinates still in pixel, but with pixels unit. world_coords_t has 1/256th pixel precision.