using cando() to handle gob/world collisions
Let's give the *Map one more interface used to test collisions between sprites and level: iWorld. The idea is to give each tile of the level some properties (solid, water, climbable, etc) that controllers can check. They do this with the help of GameObject::cando() which will evaluate one position around the current one and tell which of some interesting properties hold at that position.

Controllers, such as 'freemove' will issue several such calls adapting their desired directions to what is detected. For instance if we cannot do the diagonal (dx, dy) move, maybe we can do (dx, 0) to move horizontally only or (0, dy) to move vertially only. Or maybe we can still move (OnePixel, dy / dx * OnePixel)...

Because the 'cando' function tests whether a given property holds on all tiles of the scanned area, we won't actually use 'SOLID' as a property, but rather 'PLAYER_CAN_MOVE_THROUGH' or 'PLAYER_CAN_FALL_THROUGH' (if that's a platformer), because those properties nicely 'multiply'. As soon as one tile of the scanned area doesn't allow the player to move through, then we cannot move into that area, and cando() clears that property flag.

See http://sylvainhb.blogspot.be/2009/02/xargon-may-source-be-with-you.html for more information.
and http://sylvainhb.blogspot.be/2009/02/world-collisions.html